+++ projects/sasil-web/src/components/atoms/StyledText/StyledText.stories.tsx
@@ -2,9 +2,10 @@ import React from 'react';
 import { ComponentStory, ComponentMeta } from '@storybook/react';
 
 import COLORS from '@sasil/common/dist/constants/style';
-import { TEXT_STYLE_NAME } from '@/constants/styles/textStyles';
+import { TEXT_STYLE_NAME } from '@sasil/common/src/constants/style/textStyles';
 import StyledText, { StyledTextProps } from './StyledText';
 
+// storybook에서 color와 textStyle을 옵션으로 선택해서 테
 export default {
   title: 'components/common/atom/StyledText',
   component: StyledText,
