+++ projects/sasil-web/src/components/atoms/StyledText/index.ts
