+++ projects/sasil-web/src/constants/styles/global.css
@@ -1,0 +1,3 @@
+* {
+  font-family: Pretendard;
+}
