+++ projects/sasil-web/src/pages/_app.tsx
@@ -1,4 +1,5 @@
 import type { AppProps } from 'next/app';
+import '@/constants/styles/global.css';
 
 function MyApp({ Component, pageProps }: AppProps) {
   /* eslint-disable react/jsx-props-no-spreading */
